type problem 5

+++ src/datasets/semantic_dataset.py
@@ -48,8 +48,8 @@ class SemanticDataset(Dataset):
                 new_label[i] = np.where(label > 0, 1, 0)
         label = new_label
         image = np.expand_dims(image, axis=0)
-        image = th.from_numpy(image).to(th.float32)
-        label = th.from_numpy(label).to(th.float32)
+        image = th.from_numpy(image).to(th.float64)
+        label = th.from_numpy(label).to(th.long)
 
 
         if self.transform:
